Development

+++ src/Modules/Blog/Explorer.Blog.Core/Domain/Blog.cs
@@ -1,8 +1,12 @@
 ﻿using Explorer.BuildingBlocks.Core.Domain;
+using Explorer.Stakeholders.Core.Domain;
+
+
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
+using System.Reflection.Metadata;
 using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Linq;
@@ -15,17 +19,21 @@ namespace Explorer.Blog.Core.Domain
         Published,
         Closed,
     }
+
+    //[Table("Blogs", Schema = "blog")] 
+
     public class Blogs: Entity
     {
-        //[ForeignKey(nameof(AuthorId))]
-        public int AuthorId { get; private set; }
+        //[ForeignKey("User")]
+        public long AuthorId { get; private set; }
+        //public User User { get; set; }
         public string Title { get; private set; }
         public string Description { get; private set; }
         public DateTime CreationDate { get; private set; }
         public string? Image { get; private set; }
         public BlogStatus Status { get; private set; }
 
-        public Blogs(int authorId, string title, string description, DateTime creationDate, BlogStatus status, string? image = null)
+        public Blogs(long authorId, string title, string description, DateTime creationDate, BlogStatus status, string? image = null)
         {
             if (string.IsNullOrWhiteSpace(title)) throw new ArgumentException("Invalid title.");
             if (string.IsNullOrWhiteSpace(description)) throw new ArgumentException("Invalid description.");
